feat: refactor auth, move dir. (#2796)
* refactor auth. move dir.

* fix deploy, add README.md

* rename dir.
