+++ service/auth/deploy/clusterimage/manifests/deploy.yaml.tmpl
