+++ service/auth/deploy/manifests/configmap.yaml
@@ -1,0 +1,47 @@
+# Copyright © 2022 sealos.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+apiVersion: v1
+kind: Namespace
+metadata:
+  labels:
+    app: auth-service
+    control-plane: service-manager
+  name: auth-service-system
+---
+apiVersion: v1
+kind: ConfigMap
+metadata:
+  labels:
+    app: auth-service
+  name: auth-service-config
+  namespace: auth-service-system
+data:
+  auth.yaml: |
+    addr: :8080
+    ssoType: "casdoor"
+    callbackUrl: ""
+    oauthProviders:
+      - type: "Github"
+        clientId: ""
+        clientSecret: ""
+    ssoEndpoint: ""
+---
+apiVersion: v1
+kind: ConfigMap
+metadata:
+  name: casdoor-init-data
+  namespace: sealos
+data:
+  init_data.json: ''
